+++ examples/standalone/src/apps/Toolbox/index.ts
@@ -1,7 +1,10 @@
 // Expose only what's necessary
 
 export { I18n } from './src/i18n.js';
+export { APIAdresseGeocodingManager } from './src/lib/geocoding/APIAdresseGeocodingManager.js';
+export type { GeocodingManager } from './src/lib/geocoding/GeocodingManager.js';
 export { Manifest } from './src/manifest.js';
 export { ExportAsanaUCD } from './src/ucds/ExportAsanaUCD.js';
 export { GenerateMiscDataUCD } from './src/ucds/GenerateMiscDataUCD.js';
+export { GeocodeAddressUCD } from './src/ucds/GeocodeAddressUCD.js';
 export { PromptLLMUCD } from './src/ucds/PromptLLMUCD.js';