+++ examples/standalone/src/apps/Toolbox/src/i18n.ts
@@ -11,6 +11,8 @@ export const I18n: AppI18n = {
         uc_GenerateMiscData_i_submit_submitting: 'Generating',
         uc_GenerateMiscData_op_0_label: 'Misc',
         uc_GenerateMiscData_op_1_label: 'UUIDs',
+        uc_GeocodeAddress_i_submit_idle: 'Geocode',
+        uc_GeocodeAddress_i_submit_submitting: 'Geocoding',
         uc_PromptLLM_desc:
             'Ask something about sports to an LLM (Mistral) and get a response back (BYOT (Bring Your Own Token))',
         uc_PromptLLM_i_submit_idle: 'Send',